TST: Disable unicode test cases on Windows
Ping gh-2561

I think it makes sense to disable unicode use in the tests on windows
for now, in order to make the general problem (have it working) on
Windows more manageable. Once we have a large-ish test battery running,
we can bring them back and investigate the topic separately from
everything else that is presently going on.