RF: run: Drop support for custom saver
Th next commit will pass None as the `path` argument instead of ".",
which would be a breaking change to how the `saver` argument is.  But
adding a compatibility kludge for outside callers isn't worth the
effort.  This argument was added so that the datalad-revolution
extension could use rev_save() instead of add() as the saver, and that
code has been absorbed into the core.  So let's just ignore the
`saver` argument entirely and slate it for removal.