RF: run: Use None rather than "." to mean "save all changes"
We pass "." as the path argument to Save() to mean "save all changes".
The next commit will change the Save() call from a dataset-bound call
to an unbound call.  That means that Save(dataset=<string>, path=".")
when called in a subdirectory will just save the subdirectory changes.
Instead we need to pass path=None.