BF: run: Use unbound calls for all dataset commands
As described here [1], there are two approaches to dealing with the
current path rules: (1) resolve all paths to absolute paths and use
bound dataset methods or (2) use unbound dataset methods and relative
paths.

run() has mostly taken the first approach.  There is one spot that
incorrectly passes relative paths to install() [2].  If this is
changed to use absolute paths, it reveals a deeper incompatibility
with the new path handling on master.  rev_resolve_path() isn't happy
with absolute paths that look like "/go/../upstairs" (see its
docstring and code comments for why).  run() constructs paths like
these when the run is happening in a subdirectory and there are inputs
or outputs upstairs (see the examples in the added test).

So instead let's switch to the second approach.

Note that the compatibility kludge around the dataset handling is
mainly for the sake of datalad-htcondor.

Also note that needing to use chpwd() in several spots is ugly.  We do
this because we need to make sure we're in the correct directory when
rerunning.  We might want to make rerun() responsible for calling
chpwd() around its call of run_command(), but that would take a bit
more refactoring.

Fixes #3551.
Re: #3746

[1] https://github.com/datalad/datalad/pull/3746#issuecomment-538265382
[2] ... though these incorrect still work in at least some cases. See
    gh-3650.