ENH: Introduce datalad.log.result-level config switch
Setting this will cause messages included in command results (which are
independently rendered as part of a result) to go to the identified log
level (debug, info, warning, error). This can be used to alter the
default behavior to have 'impossible' results cause a 'warning' log
message and 'error' results yield an 'error' log message.

The default setup leads to cluttered output and double messaging.
Setting this switch to 'debug' hides all result logging and emphasizes
final results over intermediate events (where an 'error' may or may not
be an indication of a problem in the big picture of a command
execution).