TST: Mark test_run_inputs_no_annex_repo as known Windows failure
This test leads to run() calling `get(dataset=ds.path, path=['*'])`.
On non-Windows systems, the '*' is considered a literal path and
reported to be absent, as expected.  On Windows, the underlying
os.stat() call fails [0] with

    OSError: [WinError 123] The filename, directory name, or volume
    label syntax is incorrect:
    'C:\\Users\\RUNNER~1\\AppData\\Local\\Temp\\datalad_temp_5m0lh8wt\\*'

As discussed here [1], we might want to rework run() and/or
GlobbedPaths to avoid hanging on to the unmatched glob and feeding
them as a literal paths.  That might get rid of the failure for this
particular test, though it would just paper over the core Windows path
handling problem shown by this failure..

[0]: https://github.com/datalad/datalad/pull/3746/checks?check_run_id=248850583#step:8:417
[1]: https://github.com/datalad/datalad/pull/3746#issuecomment-538577417