OPT: Half the number of for-each-ref calls in GitRepo.is_with_annex()
This addresses some aspect of gh-3766

It also removes an unused feature to limit the check to remote branches.
Even the crawler doesn't use it.

It also removes an explicit check for a direct mode related branch name,
that doesn't seem to be relevant, even for direct mode repos, and
@bproldrack also had no memory on when exactly this could happen. Given
that we do not support it anymore, I decided to remove the additional
complexity.