PERF/RF: _flyweight_invalid and is_valid_*
Bind Flyweight invalidation to the instance rather than the class. This allows for is_valid_git/is_valid_annex methods, that can now reuse information stored in the object itself rather than recreating relevant path objects all the time.
In addition don't test for a non-empty .git in GitRepo.is_valid* but a particular path beneath. This is faster by a factor between 2 and 3:
%timeit path.exists() and (not path.is_dir() or head_path.exists())
4.93 µs ± 34.8 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
%timeit path.exists() and (not path.is_dir() or any(path.iterdir()))
12.8 µs ± 150 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

Finally, is_valid_annex doesn't unnecessarily do an extensive check whether annex was initialized when it's called with allow_noninitialized and furthermore comes with an option to not check validity of underlying git first. This can be used, when GitRepo's is_valid* was called before anyway.