ENH: Introduce names for candidate clone URLs
This touches upon an idea by @yaricoptic expressed in
https://github.com/datalad/datalad/issues/3827#issuecomment-545633822
and merely lays the foundation for further work.

Names are not unique (and I don't see how they could be, or what we
would gain from it), hence no dict.

Choice of names is debatable too. But given that they are not used ATM,
it doesn't really matter and can easily be changed.