track status of yielded results
Just to be sure that we don't swallow something, that we didn't intend to.
However, the only thing I can think of, is a situation where datasets we just installed are somehow screwd up
during runtime of Get and then we are acting upon them again, while working on a different input path. I have no idea how to reasonably test that.