RF: downloaders: Support non-existing directory target
As of a570fcb70 (ENH: downloaders: Ensure directories for target
exist, 2019-09-02), download() creates leading directories if it is
given a path that does not exist for _non-directory_ targets.  A
directory target is supported, but it must exist.

Move the "make directories if needed" logic early so that we can
handle directory targets as well.