DOC: download_url: Correct --dataset's description
On both master and 0.11x, there isn't an attempt to identify the
dataset from --path argument.  For example, if outside of the
</path/to/ds/> dataset, running

  $ datalad download-url --path /path/to/ds/fname
    https://www.datalad.org/img/logo/studyforrest.png

downloads the file to </path/to/ds/fname>, but it does not perform any
of the dataset-dependent functionality (e.g., saving).

Looking at 98153ecd4 (ENH: download_url: Optionally add file to
dataset, 2018-05-17), it appears that functionality was never
supported and that this description was thoughtlessly copied from an
existing --dataset description.