ENH: download_url: Use trailing separator to signal directory target
If the --path argument points to an existing directory, download_url()
will dump content to files within that directory.  The only way we
know that the user wants a directory is that one exists.  As a
consequence, if there's a typo (as described in gh-3484),
download_url() can't be aware that a directory was intended and goes
with the non-directory treatment.  When combined with --archive, this
can lead to a large number of files in a location the user didn't
intend, typically the top-level directory of the repository.

To improve this situation, require the user to tack on a trailing
separator to indicate that they want the directory treatment.  If the
user has a typo in the directory name, at least the content goes into
a misnamed subdirectory.  And because download_url() knows what the
user wanted regardless of whether the directory exists, download_url()
can now support creating directories when they don't exist, which
underneath is already supported by download().

Closes #3848.