BF: clone: Skip special remotes with no name in autoenable check
After cloning an annex repo, we go through the special remotes and
inform users about ones that are not autoenabled.  When accessing the
dict with information from git-annex:remote.log, we allow for the
special remote name to be missing.  But the downstream code can't
actually handle the None value that is used in this case, leading to
the failure seen in gh-3816.

There is nothing useful we can do without the remote name, so just
warn and move on.  As far as I'm aware, a special remote in remote.log
must have a name, with the exception of the recently added support for
"sameas" remotes.  The next commit will handle "sameas" remotes.