ENH: clone: Handle sameas special remotes in autoenable check
The last commit addressed the failure reported in gh-3816 that
happened when cloning a dataset with a sameas special remote.  However
the check is still incorrect for sameas remotes because, at the level
of .git/config, the section for a sameas remote has the sameas-uuid
value for annex-uuid, and its configuration UUID (the first field in
git-annex:remote.log) is under annex-config-uuid.  In other words,
different remote sections may share the same annex-uuid value.

Give the annex-config-uuid value precedence over annex-uuid when we
check for a remote in .git/config.  This is necessary because
otherwise we would consider a special remote enabled if _any_ of the
remotes that have the UUID are enabled.

With this change, it's likely that handling of sameas remotes across
the codebase is sufficient.  There are only two callers of
get_special_remotes(), and they both have dedicated tests for sameas
remotes.  That leaves callers of get_remotes() and
is_special_annex_remote().  Those cases rely on the remote's
.git/config section and should work transparently with sameas remotes.
There may be issues with other spots that explicitly retrieve
remote.NAME.annex-uuid, but quickly grepping through these instances,
it seems either correct or adequate to use annex-uuid rather than
annex-config-uuid for sameas remotes.

Closes #3816.