ENH: annrepo.get_special_remotes: Copy "sameas-name" to "name"
git-annex 7.20191017 added support for having special remotes that
point to the same data store [0].  In this case, the data contained in
git-annex:remote.log as returned by get_special_remotes() looks
something like this:

    {'64e9...35e3': {'sameas-name': 'sa',
                     'sameas-uuid': 'cac3...0161',
                     ...},
     'cac3...0161': {'name': 'main',
                     ...}}

The sameas entry has a "sameas-name" key rather than a "name" key, and
it has a "sameas-uuid" value that maps it to another remote.

Put the "sameas-name" under "name" so callers don't have to check both
places.  Depending on what the caller is doing, they may be able to
treat a remote the same regardless of whether it is a sameas remote.

[0]: https://git-annex.branchable.com/tips/multiple_remotes_accessing_the_same_data_store/

Re: https://github.com/datalad/datalad/pull/3856#issuecomment-554169093