RF: siblings: Do not call update() with 'path' argument
As of bc3afaee5 (RF: update() no longer uses AnnotatePaths,
2019-09-22), the path argument of update() only restricts which
subdatasets are updated when operating recursively; it doesn't specify
which top-level dataset to operate on.  And as of 02656cbcb (ENH:
Better UX in case of misspecified or pointless path constraints,
2019-09-24), update() warns if `path` is specified but `recursive` is
false.

`siblings --fetch` calls update() with a path and recursive=False, but
it wasn't adjusted for the new calling convention, so any direct
callers of `siblings --fetch` as well as create-sibling callers see a
spurious warning.

Drop the path argument from the update() call, passing the
dataset-to-update as the dataset argument.