ENH(working Draft): with explicitly empty save message, invoke editor to provide one
I do not think it is actually desired to try to make regular git
commit editor work, since we would not want it to pop up for every subdataset
we might be operating on in a recursive invocation.  With all the recursion
we better provide our own statistics and information about what is about
to happen.  In this initial commit I am not harvesting those stats, since
it would require further code RFing, and possibly some side-effect OPT to not
bother collecting and/or carrying around all the "clean" state information
about all the files in all datasets

To invoke editor upon save, just specify -m .  In the long run, we could
make it behave like git, i.e. upon not only an explicitly empty, but None
as well.