BF: Reintroduce 'notneeded' results for existing datasets in get()
However, instead of reverting to the behavior prior 0.12rc6 (blind
duplicate reporting of dataset results for subdataset install and
annex-get of its root directory, this patch compromises and actually
inspects if there already was a report on a dataset, and only silences
follow-up duplicate reports. This impacts the result composition of
large-ish operations (recursive install of a dataset with data), but
AFAICS it should have now actual impact on logic and actionability
of the results.

This fixes gh-3836