NF: Result hooks
This is aims to be a more flexible alternative to --proc-pre/post and
its proposed successor of command hooks (#3264).

The key idea is that we have DataLad's results that all pass through the
main event loop, and we can define ad-hoc hooks that run custom actions
whenever a matching result is observed.

Key differences to what we already have:

- can act more than once per command execution
- no "pre" action anymore
- informed by the actual result itself
- runs dataset procedures, but also any proper datalad commands