BF: gitrepo: Work around change in 'ls-files -o' output
save_() calls 'ls-files -o' on a list of untracked directories to
determine which directories correspond to untracked submodules.  If
the directories remain unexpanded in the 'ls-files -o' output, they
are taken as submodules.

This method of identifying untracked submodules breaks [0,1] with the
latest release of Git (v2.24.0) due to an unintentional change in the
'ls-files -o' output [2]: when given _multiple_ pathspecs, 'ls-files
-o' recurses into untracked submodules and lists the files from the
submodule (even the tracked ones).  save_() filters the output to
directories, so most of the additional entries are removed, but when
the submodule itself has submodules (untracked or tracked) these and
those from any deeper levels are included in the output.  save_()
passes these deeper repositories to add_submodule(), which as expected
leads to 'git submodule add' failing.

This behavior will hopefully be addressed in Git itself [2], but we
should still provide a workaround for the current version of Git.  Add
a helper that filters these deeper repositories from the list of
submodules that save_() feeds to add_submodule().  This approach takes
advantage of the fact that, even when 'ls-files -o' misbehaves and
traverses into the submodule, it still reports an unexpanded entry.
If the unexpanded entry weren't present, we wouldn't be able to
distinguish a submodule from a regular directory.

Note that the helper is conditionally defined at the module level; for
earlier versions of Git that don't need this kludge, this reduces the
cost per save_() call to a single call to an identity function.

[0]: https://github.com/datalad/datalad/issues/3890#issuecomment-561722194
[1]: https://github.com/datalad/datalad/pull/3902#issuecomment-562630681
[2]: https://lore.kernel.org/git/87fti15agv.fsf@kyleam.com/T/#u