TST: diff: Skip an assertion that fails with Git v2.24.0
As described in the previous commit, the traversal behavior of
'ls-files -o' in the latest release of Git, v2.24.0, has changed.  The
workaround in that commit fixed two tests that failed with v2.24.0,
but there is a remaining failure, test_path_diff [0].  Local testing
confirmed that it is due to the same change in Git, 89a1f4aaf7 (dir:
if our pathspec might match files under a dir, recurse into it,
2019-09-17).

The failures addressed in the last commit were about 'ls-files -o'
without --directory traversing into untracked nested repositories when
multiple pathspecs are given.  In contrast the test_path_diff failure
happens due to traversing into a plain untracked directory when
--directory is specified along with multiple pathspecs.  For example,
before 89a1f4aaf7 the following will not traverse into an untracked
directory:

  git-89a1f4aaf7^ $ git ls-files -o --directory . a
  a
  untracked-dir/

As of 89a1f4aaf7 the same command includes untracked content from the
untracked directory:

  git-89a1f4aaf7 $ git ls-files -o --directory . a
  a
  untracked-dir/
  untracked-dir/b

Note that if the "a" pathspec were dropped, the call would _not_
traverse into the untracked directory (i.e., "untracked-dir/b" would
not be included).

For now let's just skip the assertion on Git v2.24 because (1) unlike
the GitRepo.save_() case from the previous commit, it's not clear that
this discrepancy in the 'datalad diff' output is critical in any way,
and (2) it looks like the change in behavior may be addressed
upstream.

[0]: https://github.com/datalad/datalad/pull/3904#issuecomment-562810392