TST: Update some checks for save syncing an adjusted branch
As of a few commits back, saving an annex repo that is on an adjusted
branch calls `git annex sync` to propagate changes back to the main
line.  This causes the latest commit to be the sync commit and
confuses some checks the expect the history to be in a certain state.

The easiest confusions to address are (1) "I expect the last commit to
be this" and (2) "diffing these last commits gives ...".  Make these
work on the adjusted branch by using "master" (the main line branch)
instead of "HEAD".

Note that there are still remaining failures in listed at gh-3817.
These are more complicated because they involve subdatasets (see
gh-3818) or because the actual functionality of the commands (in
particular aggregate_metadata and rerun) was designed to work on a
non-adjusted branch.  Beyond that, only a subset of tests are executed
on Windows, those builds are the only ones where we test with adjusted
branches, so there are likely to be many more issues.

Re: https://github.com/datalad/datalad/pull/3817#issuecomment-565799813