BF: Always report absolute location of discovered procedures
Before it would report relpaths if operating on a dataset, leading to
the issue at hand.

I considered relpaths for dataset internal procedures, and abspaths for
others. However, when not invoked at the root of a dataset these would
again be somewhat confusing. So full abspath it is.

Fixes gh-3619