BF: Provide indication which procedure is (to be) execut(ed/able)
Before only "absent" (but configured) procedures were marked up
in the 'state' field of their result record. Now this is extended
with two more states:

'executable': all criteria for an active procedure are fulfilled
'overridden': like 'executable', but there is a higher-priority
              procedure with the same name

These additional states are also reflected in the renderer, where
only active/executable procedures are renderer as before, and
any other's are de-emphasized and annotated.

Fixes gh-3614