RF: Turn --reckless into a mode-option (instead of just bool)
This gives us the flexibility to support various flavors of
recklessness.

The change is implemented in a way that maintains compatiility with both
old Python and shell code. Nevertheless, all internal usage has been
ported to the new argument.

Following a suggestion in
https://github.com/datalad/datalad/issues/3957#issuecomment-568801465
the default (if enabled) is called 'auto' to make "clever" operation
the standard behavior.