NF: ConfigManager.???(where='override')
This brings the same (and actually more) configuration override
flexibility to the Python API that the cmdline API offers via
`datalad -c key=val`.

This feature is a necessity for being able to rely on one-off
configuration (changes) to alter command behavior, whenever
a dedicated command option is not absolutely needed or shall be
avoided.

A dedicated test for overrides is also included.

Fixes gh-3970