TST: create-sibling: Loosen check's assumption about Git's behavior
A check in test_target_ssh_simple assumes that executing the
post-update hook will change the modification times of .git/info/refs
and .git/objects/info/packs.  In test runs with more recent Git
versions, this check has started to fail because the modification
times of the files are not updated [0].  This change in behavior is
likely due to Git's f4f476b6a1 (update-server-info: avoid needless
overwrites, 2019-05-13), which was part of the v2.23.0 release.

We could condition the check on the Git version, but, as the now
failing check demonstrates, this check is much too concerned with
internal implementation details of git.  To avoid the test starting to
fail due to another change within .git/ that we do not control, let's
instead update the check to not assume that it knows the strict set of
files that's modified.

[0]: Re: https://github.com/datalad/datalad/issues/3890#issuecomment-565587037