RF: Attribute -> Value Error + provide an informative message for the exception
See https://github.com/datalad/datalad/issues/3973 .   I am not really sure why it was
intended to be an AttributeError, but I feel that ValueError is a better fit here.
lgr.debug message might be of help in some cases to identify the reason which lead
to crash, but because args and kwargs are modified by the code above, not sure really
if it would be very informative.  but I do not think that it would somehow negatively
effect the performance so I added it