NF: Support cloning of specific repository versions (fixes gh-2109)
`GitRepo.clone()` is enhanced to accept arbitrary additional options
that are passed on to `git-clone`. We use the `--branch` option
(that, despite its name, can handle any relevant version identifier)
to let `git-clone` do all the work.

`datalad-clone` is now set up to acknowlege a request for a particular
version, for any kind of source URL where `decode_source_spec()`
yields a non-None `version` property.

At the moment this is only the case for `ria+http|ssh://` URL, but
future additions only need to alter/enhance `decode_source_spec()`
to extend this functionality for other types.

This change also sets the stage for gh-4035