RF: Preserve filename extension in download_url(path=...) if possible
The reason is the same as what made us use MD5E instead MD5 as the
default backend. If we purposefully remove an extension, we have
mimetype inspection left is the only option to decided on some
further processing (e.g. archive type). Possible, but needlessly
complicated.

An alternative fix would have been to not purposefully use an
extension-less archive name in the associated tests.