BF: clone: Account for relative paths when inheriting local origins
As of 367454e0b (NF: Auto-configure local origin of local origin to
make annex available, 2019-12-29), we check whether origin's URL for a
newly cloned annex dataset points to a local dataset.  If it does, we
add _that_ dataset's origin (if any) to the newly cloned dataset.  And
as of d2a25d090 (ENH: Make configuration of local origin siblings work
recursively, 2020-01-05), we continue walking up the chain of local
origin's until we hit an origin that is not a local path.

However, we create the corresponding dataset instance with the wrong
path when origin is a _relative_ path: it should be relative to the
cloned repository, not the current working directory.  The incorrect
path means that there's no "origin of origin" to find, and an
"origin-N" remote isn't added.

Note that this bug isn't currently very accessible because `git clone`
converts relative paths to absolute ones, and the first attempt to
adjust these back to relative paths (02e2b4c46 and 0a80bb47e) was
reverted in e8c5d7069 (BF: clone: Revert incorrect relative path
adjustment to URLs, 2020-01-14).  But relative paths will be a common
case when the second attempt from gh-4026 lands.