NF: Intersphinx mapping to link to handbook content (fixes gh-4034)
Demo link is included. All rendered references are version to match the
current "major" release.

Looks like this:

--help output and help(clone) in Python

```
  handbook:3-001 (http://handbook.datalad.org/symbols)
    More information on Remote Indexed Archive (RIA) stores
```

The given URL points to a dedicated index section in the handbook, where
we can maintain any number of references that point to arbitrary
locations. The name 'symbols' is used, because this is how sphinx lists
numeric references. The ?-???? scheme aims to mimic manpage sections
(where we can come up with some categorization, if needed, and
incremental integer IDs).

Ping @adswa for awareness (I have pushed a corresponding change to the
handbook; directly, no PR).