BF: clone: Store stderr so it is displayed on clone failure
When GitPython raises a GitCommandError, whether the exception's
stderr attribute actually contains the standard error depends on
whether progress reporting is enabled.  With progress enabled, the
standard error is stored in RemoteProgress.error_lines instead of the
exception [0].

As of 3ecc68a82 (2018-09-27), we assign RemoteProgress.error_lines to
our GitPythonProgressBar._last_error_lines.  In the case of clone(),
we consider _last_error_lines to some extent, but it's not included in
the errors messages that we show the clone call fails with all
candidates.  In the specific cases of an unknown version, this means
that we report that cloning failed without any indication that the
unknown version was the issue.

Make the standard error visible to the caller by replacing the
exception's stderr with _last_error_lines.

Fixes #4038.

[0]: https://github.com/gitpython-developers/GitPython/issues/599