BF: external_versions: Drop exception handling from ssh/7z methods
When _get_system_ssh_version() and _get_system_7z_version() call their
commands through the Runner, they catch a CommandError, log it, and
return None.  This is unnecessary because
ExternalVersions.__getitem__() already catches and logs all exceptions
when it calls the custom _get*version() functions.

Also, probably the most common reason the command would fail is that
the program itself is not found.  But these command calls use the list
form, and in that case subprocess will raise a FileNotFoundError, not
a CommandError.

    $ python -c "from datalad.cmd import Runner; Runner()(['not', 'there'])"
    Traceback (most recent call last):
      File "<string>", line 1, in <module>
      File "/[...]/datalad/datalad/cmd.py", line 185, in __call__
        return self.run(cmd, *args, **kwargs)
      File "/[...]/datalad/datalad/cmd.py", line 500, in run
        stdin=stdin)
      File "/usr/lib/python3.7/subprocess.py", line 775, in __init__
        restore_signals, start_new_session)
      File "/usr/lib/python3.7/subprocess.py", line 1522, in _execute_child
        raise child_exception_type(errno_num, err_msg, err_filename)
    FileNotFoundError: [Errno 2] No such file or directory: 'not': 'not'