NF: Command abstractions
This is more thinking out loud than a concrete implementation. The idea
is this:

Let's collect basic operation into classes that can take care of doing
the right thing under different conditions in order to achieve this.
At the moment, we have the Runner to help with that, but the actual
nature of the operations done are scattered throughout the code.

This limits us, and, for example, is something that contributes to the
fact that something like `create_sibling()` can only work with
POSIX-like system, via SSH connections. Moving the local and remote
operation it performs into helper classes that can be specialized for any
target scenario, would help simplify the implementation of that command,
while simultaneously make it more flexible, e.g. it could discover
that the target sibling is at a local path, and internally pick a local
operations helper and achieve the desired result without having to go
through SSH.

Likewise, any, now hardcoded, command execution could be centrally
switched to a variant that is optimized for the local platform
(look for `LocalOperations` in the diff).

Another angle for mid-term improvements would be that less code talks
directly to a Runner. Having specialized methods for particular
operations would enable us to optimize the execution to meet our needs
better.

I envision no limit to the scope of supported operations. For example,
something like `enable_post_update_hook(path)` would be perfectly
in-scope. But even `edit_textfile_interactively(path)` is easily within
the range of possibilities. It would be important, though, that the
classes remain cheap to instantiate.

It is also fine, if not all operations are implemented for all potential
target scenarios immediately. However, the API of an operation would
need to be carefully crafted to avoid overfitting any particular
environment or use case.

We should also consider a similar approach for interfacing with Git
commands, in particular those that report paths -- which are reported in
POSIX conventions, but were/are inappropriately treated as native paths,
leading to failure on Windows. However, it is less clear to me how to
consolidate this with our current approach (ie.  GitRepo.call_git...()).

In conclusion, I'd argue that having such an abstraction layer would
benefit old code by making it more flexible with little effect, and aid
the design of new code that could be written with the fact in mind that
low-level operations are taken care off in a platform-appropriate
fashion, but also that they have to be implemented in a
platform-appropriate fashion, and where this can be done.

[skip ci]