BF: Prevent accidential read from AND write to undesired datasets
The cause of gh-4071 are call to `git-config` typically using
PWD as CWD of the runner. This made us read and write to whatever
repository was at PWD instead of a dataset a config manager was
parametrized with, when using `where=local' (which we don't do often
outside the tests, but increasingly do).

The fix is twofold:

- give an appropriate CWD to the Runner, based on a per-call
  evaluation whether the dataset path actually exists.

- more comprehensive evaluation of contractor and method parameters
  and exceptions in case of invalid calls (e.g. modify local config,
  without a dataset being given, etc.)

Given that this is a critical piece of code that runs a lot, any such
addition check hurts (performance improvements are much welcome!), but
given that we were doing this so wrong for so long send a clear message
what type of compromise to prefer...

Fixes gh-4071