BF: Make sure source candidate configs for submodules are inherited
If they do not come from system config.

This makes it possible to have large collections of datasets put in
some dataset store, without any of them carrying knowledge about their
hosting location, even if they have super-sub-dataset relationships
among themselves. The only requirement to make any of these datasets
accessible, is a single top-level superdataset with a source candidate
configuration that is point to the store. Everything else will be
handle with on-clone configuration in `.git/config` of each subdataset.